added paths; replaced DOUBLE with DECIMAL for the balance field in the db & model

+++ src/main/java/com/springboot/simpleatm/model/Account.java
@@ -5,6 +5,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Entity
 @Getter @Setter @NoArgsConstructor
@@ -17,9 +18,9 @@ public class Account {
     private String cardNumber;
     @Column(name = "pin_hash")
     private String pinHash;
-    private Double balance;
+    private BigDecimal balance;
 
-    public Account(String name, String cardNumber, String pinHash, Double balance) {
+    public Account(String name, String cardNumber, String pinHash, BigDecimal balance) {
         this.name = name;
         this.cardNumber = cardNumber;
         this.pinHash = pinHash;
